Move service executors to runtime_local module
