+++ libs/executors/CMakeLists.txt
@@ -46,7 +46,6 @@ set(executors_compat_headers
     hpx/execution/executors/parallel_executor.hpp
     hpx/execution/executors/restricted_thread_pool_executor.hpp
     hpx/execution/executors/sequenced_executor.hpp
-    hpx/execution/executors/service_executors.hpp
     hpx/execution/executors/thread_pool_attached_executors.hpp
     hpx/parallel/datapar/execution_policy_fwd.hpp
     hpx/parallel/datapar/execution_policy.hpp
@@ -57,7 +56,6 @@ set(executors_compat_headers
     hpx/parallel/executors/parallel_executor_aggregated.hpp
     hpx/parallel/executors/parallel_executor.hpp
     hpx/parallel/executors/sequenced_executor.hpp
-    hpx/parallel/executors/service_executors.hpp
     hpx/parallel/executors/thread_pool_attached_executors.hpp
 )
