+++ libs/executors/tests/unit/service_executors.cpp
@@ -4,9 +4,9 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#include <hpx/executors/service_executors.hpp>
 #include <hpx/hpx.hpp>
 #include <hpx/hpx_init.hpp>
+#include <hpx/include/thread_executors.hpp>
 #include <hpx/testing.hpp>
 
 #include <algorithm>
