+++ libs/runtime_local/include/hpx/runtime_local/run_as_os_thread.hpp
@@ -9,10 +9,10 @@
 #include <hpx/config.hpp>
 #include <hpx/assertion.hpp>
 #include <hpx/execution/executors/execution.hpp>
-#include <hpx/executors/service_executors.hpp>
 #include <hpx/functional/deferred_call.hpp>
 #include <hpx/functional/result_of.hpp>
 #include <hpx/runtime/threads_fwd.hpp>
+#include <hpx/runtime_local/service_executors.hpp>
 
 #include <type_traits>
 #include <utility>
