+++ libs/runtime_local/include_compatibility/hpx/parallel/executors/service_executors.hpp
@@ -1,4 +1,4 @@
-//  Copyright (c) 2019 Ste||ar Group
+//  Copyright (c) 2020 ETH Zurich
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
@@ -7,17 +7,17 @@
 #pragma once
 
 #include <hpx/config.hpp>
-#include <hpx/executors/config/defines.hpp>
-#include <hpx/executors/executors/service_executors.hpp>
+#include <hpx/runtime_local/config/defines.hpp>
+#include <hpx/runtime_local/service_executors.hpp>
 
-#if defined(HPX_EXECUTION_HAVE_DEPRECATION_WARNINGS)
+#if defined(HPX_RUNTIME_LOCAL_HAVE_DEPRECATION_WARNINGS)
 #if defined(HPX_MSVC)
 #pragma message(                                                               \
     "The header hpx/parallel/executors/service_executors.hpp is deprecated, \
-    please include hpx/executors/service_executors.hpp instead")
+    please include hpx/runtime_local/service_executors.hpp instead")
 #else
 #warning                                                                       \
     "The header hpx/parallel/executors/service_executors.hpp is deprecated, \
-    please include hpx/executors/service_executors.hpp instead"
+    please include hpx/runtime_local/service_executors.hpp instead"
 #endif
 #endif
