+++ libs/thread_executors/CMakeLists.txt
@@ -22,7 +22,6 @@ set(thread_executors_headers
     hpx/thread_executors/limiting_executor.hpp
     hpx/thread_executors/manage_thread_executor.hpp
     hpx/thread_executors/resource_manager.hpp
-    hpx/thread_executors/service_executors.hpp
     hpx/thread_executors/this_thread_executors.hpp
     hpx/thread_executors/thread_execution.hpp
     hpx/thread_executors/thread_execution_information.hpp
@@ -46,7 +45,6 @@ set(thread_executors_compat_headers
     hpx/runtime/threads/executors/guided_pool_executor.hpp
     hpx/runtime/threads/executors/limiting_executor.hpp
     hpx/runtime/threads/executors/manage_thread_executor.hpp
-    hpx/runtime/threads/executors/service_executors.hpp
     hpx/runtime/threads/executors/this_thread_executors.hpp
     hpx/runtime/threads/executors/thread_pool_attached_executors.hpp
     hpx/runtime/threads/executors/thread_pool_os_executors.hpp
