--- libs/thread_executors/include/hpx/thread_executors/service_executors.hpp
@@ -1,24 +1,0 @@
-//  Copyright (c) 2007-2016 Hartmut Kaiser
-//
-//  SPDX-License-Identifier: BSL-1.0
-//  Distributed under the Boost Software License, Version 1.0. (See accompanying
-//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-
-#pragma once
-
-#include <hpx/config.hpp>
-
-#if defined(HPX_HAVE_THREAD_EXECUTORS_COMPATIBILITY)
-#include <hpx/executors/service_executors.hpp>
-
-namespace hpx { namespace threads { namespace executors {
-    using parallel::execution::service_executor_type;
-
-    using parallel::execution::io_pool_executor;
-    using parallel::execution::main_pool_executor;
-    using parallel::execution::parcel_pool_executor;
-    using parallel::execution::service_executor;
-    using parallel::execution::timer_pool_executor;
-}}}    // namespace hpx::threads::executors
-
-#endif
